Moved all databaseIO tests to the main test file and made sure they all work. Removed the removeItem call in the UserIO file to stop the method from deleting a furniture item every time it is called, and affecting tests.
