+++ src/edu/ucalgary/ensf409/SupplyChainTest.java
@@ -25,8 +25,11 @@ import java.util.Random;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import static junit.framework.TestCase.*;
-import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.*;
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
 
 
 /**
@@ -36,6 +39,9 @@ import static org.junit.Assert.assertFalse;
 public class SupplyChainTest {
     private final ByteArrayOutputStream terminalContent
             = new ByteArrayOutputStream();
+    private final ByteArrayOutputStream errContent
+            = new ByteArrayOutputStream();
+    private final PrintStream originalErrStream = System.err;
     private final PrintStream originalTermContent = System.out;
     private final InputStream originalTerminalInput = System.in;
 
@@ -50,6 +56,8 @@ public class SupplyChainTest {
     public void readTerminalOutputSetup() {
         System.setOut(new PrintStream(terminalContent));
         System.setIn(originalTerminalInput);
+        System.setErr(new PrintStream(errContent));
+
     }
 
     /**
@@ -59,6 +67,7 @@ public class SupplyChainTest {
     @After
     public void restoreTerminalOutputStream() {
         System.setOut(originalTermContent);
+        System.setErr(originalErrStream);
     }
 
 
@@ -958,18 +967,68 @@ public class SupplyChainTest {
         assertTrue(fileUpdated && correctOutput);
     }
 
-    /**
-     * The alteration of the MySQL credentials cannot be tested reasonably
-     * without knowledge of all other accounts on the system Supply Chain
-     * Manager is being run on because the test may be rendered incorrectly if
-     * the values used for testing to not represent an existing MySQL database
-     * and respective user does not exist on the current machine.
-     *
-     * The choice to have the database credentials' validity tested immediately
-     * after the user has changed them intends to alert the user of a problem
-     * with their current credentials before they continue to program and
-     * attempt to access and manipulate data they do not have access to.
-     */
+    @Test public void testUserIO_inValidCredentialsChange(){
+        ByteArrayInputStream terminalInput1 =
+                new ByteArrayInputStream(("\nY\n" +
+                        "invalidURL\n" +
+                        "invalidUsername\n" +
+                        "invalidPassword\n" +
+                        "Y\n\n").getBytes());
+        System.setIn(terminalInput1);
+        UserIO userIO = new UserIO();
+
+
+        userIO.processInput(5); // Processes the input for
+        // requesting to change the
+        //current MySQL credentials to invalid values
+
+        userIO.processInput(1); // Process the input for requesting
+        // a new order using the invalid MySQL Credentials
+
+
+        String expectedOutputREGEX = "(Unable to create a connection with\\Rthe credentials:\\R[ ]{5}DbURL: invalidURL\\R[ ]{2}Username: invalidUsername\\R[ ]{2}Password: invalidPassword)";
+        Pattern expectedPattern = Pattern.compile(expectedOutputREGEX);
+        Matcher expectedMatch = expectedPattern.matcher
+                (terminalContent.toString());
+
+
+        boolean correctOutput = expectedMatch.find();
+        assertTrue( correctOutput);
+    }
+
+    @Test public void testUserIO_ValidCredentialsChange(){
+        ByteArrayInputStream terminalInput1 =
+                new ByteArrayInputStream(("\nY\n" +
+                        "invalidURL\n" +
+                        "invalidUsername\n" +
+                        "invalidPassword\n" +
+                        "Y\n\n" +
+                        "\nY\n" +
+                        "jdbc:mysql://localhost/inventory\n" +
+                        "scm\n" +
+                        "ensf409\n" +
+                        "Y\n\nCANCEL").getBytes());
+        System.setIn(terminalInput1);
+        UserIO userIO = new UserIO();
+        userIO.processInput(5); // Processes the input for
+        // requesting to change the
+        // current MySQL credentials to invalid values.
+        userIO.processInput(5); // Processes the input for
+        // requesting to change the
+        //current MySQL credentials to valid values.
+
+        userIO.processInput(1); // Process the input for requesting
+        // a new order using the valid MySQL Credentials
+
+        String expectedOutputREGEX = "(Please input request for furniture item)";
+        Pattern expectedPattern = Pattern.compile(expectedOutputREGEX);
+        Matcher expectedMatch = expectedPattern.matcher
+                (terminalContent.toString());
+
+
+        boolean correctOutput = expectedMatch.find();
+        assertTrue( correctOutput);
+    }
 
 
     /**
@@ -997,4 +1056,377 @@ public class SupplyChainTest {
         assertTrue(fromOption2 && fromOption3);
     }
 
-}
+
+
+        /**
+         * testDatabaseIO_testValidChairDataTypeTask asserts that the
+         * PriceOptimizer object created by the method getChairData produces
+         * the correct String array of ID's, int array of prices, and boolean
+         * 2D array of parts for the type "Task"
+         */
+        @Test
+        public void testDatabaseIO_testValidChairDataTypeTask(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            String[] id = {"C0914", "C1148", "C3405"};
+            int[] price = {50, 125, 100};
+            boolean [][] parts = {{false, false, true, true},
+                    {true, false, true, true} ,{true, true, false, false}};
+
+            PriceOptimizer arrays = database.getChairData("Task");
+
+            assertTrue(Arrays.equals(id, arrays.getId())
+                    && Arrays.equals(price, arrays.getPrice())
+                    && Arrays.deepEquals(parts, arrays.getParts()));
+
+        }
+
+        /**
+         * testDatabaseIO_testValidChairDataTypeMesh asserts that the
+         * PriceOptimizer object created by the method getChairData produces
+         * the correct String array of ID's, int array of prices, and boolean
+         * 2D array of parts for the type "Mesh"
+         */
+        @Test
+        public void testDatabaseIO_testValidChairDataTypeMesh(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            String[] id = {"C0942", "C6748", "C8138", "C9890"};
+            int[] price = {175, 75, 75, 50};
+            boolean [][] parts = {{true, false, true, true},
+                    {true, false, false, false} ,{false, false, true, false},
+                    {false, true, false, true}};
+
+            PriceOptimizer arrays = database.getChairData("Mesh");
+
+            assertTrue(Arrays.equals(id, arrays.getId())
+                    && Arrays.equals(price, arrays.getPrice())
+                    && Arrays.deepEquals(parts, arrays.getParts()));
+        }
+
+        /**
+         * testDatabaseIO_testInvalidChairData asserts that if an
+         * invalid type is entered into the method getLampData,
+         * the PriceOptimizer object created is null.
+         */
+        @Test
+        public void testDatabaseIO_testInvalidChairData(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            assertNull(database.getChairData("invalid"));
+        }
+
+        /**
+         * testDatabaseIO_testValidDeskDataTypeStanding asserts that the
+         * PriceOptimizer object created by the method getDeskData produces
+         * the correct String array of ID's, int array of prices, and boolean
+         * 2D array of parts for the type "Standing"
+         */
+        @Test
+        public void testDatabaseIO_testValidDeskDataTypeStanding(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            String[] id = {"D1927", "D2341", "D3820", "D4438", "D9387"};
+            int[] price = {200, 100, 150, 150, 250};
+            boolean [][] parts = {{true, false, true}, {false, true, false}
+                    ,{true, false, false}, {false, true, true},
+                    {true, true, false}};
+
+            PriceOptimizer arrays = database.getDeskData("Standing");
+
+            assertTrue(Arrays.equals(id, arrays.getId())
+                    && Arrays.equals(price, arrays.getPrice())
+                    && Arrays.deepEquals(parts, arrays.getParts()));
+        }
+
+        /**
+         * testDatabaseIO_testValidDeskDataTypeTraditional asserts that the
+         * PriceOptimizer object created by the method getDeskData produces
+         * the correct String array of ID's, int array of prices, and boolean
+         * 2D array of parts for the type "Traditional"
+         */
+        @Test
+        public void testDatabaseIO_testValidDeskDataTypeTraditional(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            String[] id = {"D0890", "D4231", "D8675", "D9352"};
+            int[] price = {25, 50, 75, 75};
+            boolean [][] parts = {{false, false, true}, {false, true, true}
+                    ,{true, true, false}, {true, false, true}};
+
+            PriceOptimizer arrays = database.getDeskData("Traditional");
+
+            assertTrue(Arrays.equals(id, arrays.getId())
+                    && Arrays.equals(price, arrays.getPrice())
+                    && Arrays.deepEquals(parts, arrays.getParts()));
+        }
+
+        /**
+         * testDatabaseIO_testInvalidDeskData asserts that if an
+         * invalid type is entered into the method getDeskData,
+         * the PriceOptimizer container created is null.
+         */
+        @Test
+        public void testDatabaseIO_testInvalidDeskData(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            assertNull(database.getDeskData("invalid"));
+        }
+
+        /**
+         * testDatabaseIO_testValidLampDataTypeSwingArm asserts that the
+         * PriceOptimizer container created by the method getLampData produces
+         * the correct String array of ID's, int array of prices, and boolean
+         * 2D array of parts for the type "Swing Arm"
+         */
+        @Test
+        public void testDatabaseIO_testValidLampDataTypeSwingArm(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            String[] id = {"L053", "L096", "L487", "L879"};
+            int[] price = {27, 3, 27, 3};
+            boolean [][] parts = {{true, false}, {false, true}
+                    ,{true, false}, {false, true}};
+
+            PriceOptimizer arrays = database.getLampData("Swing Arm");
+
+            assertTrue(Arrays.equals(id, arrays.getId())
+                    && Arrays.equals(price, arrays.getPrice())
+                    && Arrays.deepEquals(parts, arrays.getParts()));
+        }
+
+        /**
+         * testDatabaseIO_testValidLampDataTypeStudy asserts that the
+         * PriceOptimizer container created by the method getLampData produces
+         * the correct String array of ID's, int array of prices, and boolean
+         * 2D array of parts for the type "Study"
+         */
+        @Test
+        public void testDatabaseIO_testValidLampDataTypeStudy(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            String[] id = {"L223", "L928", "L980", "L982"};
+            int[] price = {2, 10, 2, 8};
+            boolean [][] parts = {{false, true}, {true, true}
+                    ,{false, true}, {true, false}};
+
+            PriceOptimizer arrays = database.getLampData("Study");
+
+            assertTrue(Arrays.equals(id, arrays.getId())
+                    && Arrays.equals(price, arrays.getPrice())
+                    && Arrays.deepEquals(parts, arrays.getParts()));
+        }
+
+        /**
+         * testDatabaseIO_testInvalidLampData asserts that if an
+         * invalid type is entered into the method getLampData,
+         * the PriceOptimizer container created is null.
+         */
+        @Test
+        public void testDatabaseIO_testInvalidLampData(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            assertNull(database.getLampData("invalid"));
+        }
+
+        /**
+         * testDatabaseIO_testValidFilingDataTypeLarge asserts that the
+         * PriceOptimizer container created by the method getFilingData produces
+         * the correct String array of ID's, int array of prices, and boolean
+         * 2D array of parts for the type "Large"
+         */
+        @Test
+        public void testDatabaseIO_testValidFilingDataTypeLarge(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            String[] id = {"F003", "F010", "F011", "F012", "F015"};
+            int[] price = {150, 225, 225, 75, 75};
+            boolean [][] parts = {{false, false, true}, {true, false, true}
+                    ,{false, true, true}, {false, true, false},
+                    {true, false, false}};
+
+            PriceOptimizer arrays = database.getFilingData("Large");
+
+            assertTrue(Arrays.equals(id, arrays.getId())
+                    && Arrays.equals(price, arrays.getPrice())
+                    && Arrays.deepEquals(parts, arrays.getParts()));
+        }
+
+        /**
+         * testDatabaseIO_testValidFilingDataTypeSmall asserts that the
+         * PriceOptimizer container created by the method getFilingData produces
+         * the correct String array of ID's, int array of prices, and boolean
+         * 2D array of parts for the type "Small"
+         */
+        @Test
+        public void testDatabaseIO_testValidFilingDataTypeSmall(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            String[] id = {"F001", "F004", "F005", "F006", "F013"};
+            int[] price = {50, 75, 75, 50, 50};
+            boolean [][] parts = {{true, true, false}, {false, true, true}
+                    ,{true, false, true}, {true, true, false},
+                    {false, false, true}};
+
+            PriceOptimizer arrays = database.getFilingData("Small");
+
+            assertTrue(Arrays.equals(id, arrays.getId())
+                    && Arrays.equals(price, arrays.getPrice())
+                    && Arrays.deepEquals(parts, arrays.getParts()));
+        }
+
+        /**
+         * testDatabaseIO_testInvalidFilingData asserts that if an
+         * invalid type is entered into the method getLampData,
+         * the PriceOptimizer container created is null.
+         */
+        @Test
+        public void testDatabaseIO_testInvalidFilingData(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            assertNull(database.getFilingData("invalid"));
+        }
+
+        /**
+         * testDatabaseIO_testSuggestedManufacturersChair asserts
+         * that the String ArrayList created by the method suggestedManufacturers
+         * produces the correct String ArrayList of the suggested
+         * manufacturers for chairs.
+         */
+        @Test
+        public void testDatabaseIO_testSuggestedManufacturersChair(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            String[] checkChair = {"Office Furnishings", "Chairs R Us",
+                    "Furniture Goods", "Fine Office Supplies"};
+            ArrayList<String> test = database.suggestedManufacturers("chair");
+            String[] testArray = test.toArray(new String[0]);
+
+            assertArrayEquals(checkChair, testArray);
+        }
+
+        /**
+         * testDatabaseIO_testSuggestedManufacturersLamp asserts
+         * that the String ArrayList created by the method suggestedManufacturers
+         * produces the correct String ArrayList of the suggested
+         * manufacturers for lamps.
+         */
+        @Test
+        public void testDatabaseIO_testSuggestedManufacturersLamp(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            String[] checkChair = {"Office Furnishings", "Furniture Goods",
+                    "Fine Office Supplies"};
+            ArrayList<String> test = database.suggestedManufacturers("lamp");
+            String[] testArray = test.toArray(new String[0]);
+
+            assertArrayEquals(checkChair, testArray);
+        }
+
+        /**
+         * testDatabaseIO_testSuggestedManufacturersInvalid asserts
+         * that if an invalid furniture is entered into the method
+         * suggestedManufacturers, the String Arraylist created is null.
+         */
+        @Test
+        public void testDatabaseIO_testSuggestedManufacturersInvalid(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+
+            assertNull(database.suggestedManufacturers("invalid"));
+        }
+
+        /**
+         * testDatabaseIO_testGetSizeOfDeskTypeStanding asserts
+         * that the int created by the method getSize produces
+         * the correct int of the number of standing desks.
+         */
+        @Test
+        public void testDatabaseIO_testGetSizeOfDeskTypeStanding(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            assertEquals(5, database.getSize("desk", "Standing"));
+        }
+
+        /**
+         * testDatabaseIO_testGetSizeOfLampTypeDesk asserts
+         * that the int created by the method getSize produces
+         * the correct int of the number of desk lamps.
+         */
+        @Test
+        public void testDatabaseIO_testGetSizeOfLampTypeDesk(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            assertEquals(7, database.getSize("lamp", "Desk"));
+        }
+
+        /**
+         * testDatabaseIO_testGetSizeOfFilingInvalidType asserts
+         * that if an invalid type is entered into the method getSize,
+         * the int created is 0.
+         */
+        @Test
+        public void testDatabaseIO_testGetSizeOfFilingInvalidType(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            assertEquals(0, database.getSize("filing", "invalid"));
+        }
+
+        /**
+         * testDatabaseIO_testTypeMediumExistsOfFiling asserts that
+         * the correct table name "Filing" and furniture type "Medium"
+         * is entered into the method typeExists.
+         */
+        @Test
+        public void testDatabaseIO_testTypeMediumExistsOfFiling(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            assertTrue(database.typeExists("filing", "Medium"));
+        }
+
+        /**
+         * testDatabaseIO_testTypeAdjustableOfDesk asserts that
+         * the correct table name "Desk" and furniture type "Adjustable"
+         * is entered into the method typeExists.
+         */
+        @Test
+        public void testDatabaseIO_testTypeAdjustableOfDesk(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            assertTrue(database.typeExists("desk", "Adjustable"));
+        }
+
+        /**
+         * testDatabaseIO_testTypeInvalidOfChair asserts that
+         * the correct table name "Chair" but incorrect furniture
+         * type "invalid" is entered into the method typeExists.
+         */
+        @Test
+        public void testDatabaseIO_testTypeInvalidOfChair(){
+            DatabaseIO database = new DatabaseIO();
+            database.createConnection();
+
+            assertFalse(database.typeExists("chair", "invalid"));
+        }
+    }
+
