+++ src/edu/ucalgary/ensf409/UserIO.java
@@ -164,14 +164,13 @@ public class UserIO {
         switch(inputValue){
             case 1:
                 setReadValuesNull();
-                readLine();
+
                boolean connectionMade = databaseIO.createConnection();
                if(!connectionMade) {
                    System.out.println("Returning to menu. " +
                            "No order has been placed.");
                    return true;
                }
-
                 System.out.println("\nPlease input request for " +
                         "furniture item in the form");
                 System.out.println("[type] [furniture category], " +
@@ -179,6 +178,7 @@ public class UserIO {
                 System.out.println("ex. Mesh chair, 1");
                 System.out.println("or enter \"CANCEL\" to return to the menu");
                 System.out.println("Enter request: ");
+                readLine();
                 String readFromScan = readLine();
                 processUserRequest(readFromScan);
                 return true;
@@ -337,7 +337,7 @@ public class UserIO {
                             .optimize(Integer.parseInt(numOfItems));
                     if(temp!=null){
                         for(String t: temp){
-                            databaseIO.removeItem(furnCategory, t);
+                            //databaseIO.removeItem(furnCategory, t);
                         }
                         fileIO = new FileIO(outputFile,temp,userRequest,
                                 priceOpt.getCurrentCost());
